Debug auto commit
Previous two commits didn't work as
expected

It looks like it might be related to
file locking, so this time I closed
editor on my PC

This wouldn't normally be a problem.
It's happening because I am trying to
be too cute by committing my non-Atari
code changes from the Atari.

This commit was made from from my Atari 600XL!!